Finish moving code to @packages/shared
