+++ apps/api/tsconfig.json
@@ -14,9 +14,5 @@
     "types": ["bun-types"]
   },
   "exclude": ["dist", "node_modules"],
-  "include": [
-    "src",
-    // Include the Pendulum types so we can use them in our code
-    "node_modules/@pendulum-chain/types"
-  ]
+  "include": ["src", "node_modules/@pendulum-chain/types"]
 }
