+++ apps/rebalancer/package.json
@@ -8,8 +8,7 @@
     "big.js": "^7.0.1",
     "graphql": "~16.9.0",
     "graphql-request": "~6.1.0",
-    "viem": "^2.21.3",
-    "vortex-backend": "workspace:*"
+    "viem": "^2.21.3"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^5.0.2",
