+++ packages/shared/src/services/nabla/transactions/index.ts
@@ -1,6 +1,5 @@
-import { AccountMeta, encodeSubmittableExtrinsic, Networks, PendulumTokenDetails } from "@packages/shared";
+import { AccountMeta, ApiManager, encodeSubmittableExtrinsic, Networks, PendulumTokenDetails } from "@packages/shared";
 import { CreateExecuteMessageExtrinsicOptions } from "@pendulum-chain/api-solang";
-import { ApiManager } from "../../pendulum/apiManager";
 import { prepareNablaApproveTransaction } from "./approve";
 import { prepareNablaSwapTransaction } from "./swap";
