+++ packages/shared/src/services/squidrouter/config.ts
