+++ packages/shared/src/services/squidrouter/index.ts
@@ -1,1 +1,6 @@
 export * from "./axelar";
+export * from "./config";
+export * from "./offramp";
+export * from "./onramp";
+export * from "./payload";
+export * from "./route";
