+++ packages/shared/src/services/squidrouter/offramp.ts
@@ -3,8 +3,8 @@ import { u8aToHex } from "@polkadot/util";
 import { decodeAddress } from "@polkadot/util-crypto";
 import { createPublicClient, encodeFunctionData, http } from "viem";
 import { moonbeam } from "viem/chains";
-import erc20ABI from "../../../../contracts/ERC20";
-import { createRandomString, createSquidRouterHash } from "../../../helpers/squidrouter";
+import erc20ABI from "../../contracts/ERC20";
+import { createRandomString, createSquidRouterHash } from "../../helpers/squidrouter";
 import { getSquidRouterConfig } from "./config";
 import encodePayload from "./payload";
 import { createOfframpRouteParams, getRoute } from "./route";
