+++ packages/shared/src/services/xcm/index.ts
@@ -1,0 +1,5 @@
+export * from "./assethubToPendulum";
+export * from "./moonbeamToPendulum";
+export * from "./pendulumToAssethub";
+export * from "./pendulumToMoonbeam";
+export * from "./send";
