+++ packages/shared/src/services/xcm/pendulumToAssethub.ts
@@ -1,9 +1,8 @@
-import { PendulumCurrencyId } from "@packages/shared";
+import { ApiManager, PendulumCurrencyId } from "@packages/shared";
 import { SubmittableExtrinsic } from "@polkadot/api-base/types";
 import { ISubmittableResult } from "@polkadot/types/types";
 import { u8aToHex } from "@polkadot/util";
 import { decodeAddress } from "@polkadot/util-crypto";
-import { ApiManager } from "../../pendulum/apiManager";
 
 export async function createPendulumToAssethubTransfer(
   destinationAddress: string,
